requirements.txt
flask
pymupdf
google-generativeai
python-dotenv